+++ src/main/java/org/saturnclient/ui2/screens/ShiftMenu.java
@@ -3,7 +3,7 @@ package org.saturnclient.ui2.screens;
 import org.saturnclient.saturnclient.SaturnClient;
 import org.saturnclient.saturnclient.SaturnClientConfig;
 import org.saturnclient.saturnclient.menus.HudEditor;
-import org.saturnclient.ui.Textures;
+import org.saturnclient.ui2.resources.Textures;
 import org.saturnclient.ui2.Element;
 import org.saturnclient.ui2.SaturnScreen;
 import org.saturnclient.ui2.elements.Button;
